phase_3/mcp/requirements.txt
openai==1.3.7
pydantic==2.5.0
python-dotenv==1.0.0
uvicorn==0.24.0
fastapi==0.104.1
sse-starlette==1.6.5
pydantic-settings==2.1.0